Conflicts:
	qa/integration-tests-engine/pom.xml

+++ engine/src/test/java/org/camunda/bpm/engine/test/cmmn/decisiontask/DmnDecisionTaskResultListenerTest.java
@@ -1,8 +1,12 @@
-/* Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -116,7 +120,7 @@ public class DmnDecisionTaskResultListenerTest extends CmmnProcessEngineTestCase
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(0, firstOutput.getFirstEntry());
+    assertEquals(0, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(0), firstOutput.getFirstEntryTyped());
   }
 
@@ -134,7 +138,7 @@ public class DmnDecisionTaskResultListenerTest extends CmmnProcessEngineTestCase
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(12, firstOutput.getFirstEntry());
+    assertEquals(12, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(12), firstOutput.getFirstEntryTyped());
   }
 
@@ -145,7 +149,7 @@ public class DmnDecisionTaskResultListenerTest extends CmmnProcessEngineTestCase
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(33, firstOutput.getFirstEntry());
+    assertEquals(33, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(33), firstOutput.getFirstEntryTyped());
   }
