+++ engine/src/test/java/org/camunda/bpm/engine/test/dmn/businessruletask/DmnDecisionResultListenerTest.java
@@ -1,5 +1,9 @@
-/* Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
@@ -10,7 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.camunda.bpm.engine.test.dmn.businessruletask;
 
 import java.util.Collections;
@@ -120,7 +123,7 @@ public class DmnDecisionResultListenerTest extends PluggableProcessEngineTestCas
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(0, firstOutput.getFirstEntry());
+    assertEquals(0, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(0), firstOutput.getFirstEntryTyped());
   }
 
@@ -138,7 +141,7 @@ public class DmnDecisionResultListenerTest extends PluggableProcessEngineTestCas
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(12, firstOutput.getFirstEntry());
+    assertEquals(12, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(12), firstOutput.getFirstEntryTyped());
   }
 
@@ -149,7 +152,7 @@ public class DmnDecisionResultListenerTest extends PluggableProcessEngineTestCas
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(33, firstOutput.getFirstEntry());
+    assertEquals(33, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(33), firstOutput.getFirstEntryTyped());
   }
